requirements.txt
colorama==0.4.6
PyYAML==6.0.3
tqdm==4.67.3
